+++ forui/test/src/widgets/header/header_action_test.dart
@@ -7,14 +7,14 @@ import '../../test_scaffold.dart';
 
 void main() {
   testWidgets('onStateChange & onHoverChange callback called', (tester) async {
-    Set<WidgetState>? states;
+    FWidgetStatesDelta? delta;
     bool? hovered;
     await tester.pumpWidget(
       TestScaffold(
         child: FHeader.nested(
           title: const Text('Title'),
           prefixes: [
-            FHeaderAction.back(onHoverChange: (v) => hovered = v, onStateChange: (v) => states = v, onPress: () {}),
+            FHeaderAction.back(onHoverChange: (v) => hovered = v, onStateChange: (v) => delta = v, onPress: () {}),
           ],
         ),
       ),
@@ -26,13 +26,13 @@ void main() {
     await gesture.moveTo(tester.getCenter(find.byType(FHeaderAction)));
     await tester.pumpAndSettle();
 
-    expect(states, {WidgetState.hovered});
+    expect(delta, FWidgetStatesDelta({}, {WidgetState.hovered}));
     expect(hovered, true);
 
     await gesture.moveTo(Offset.zero);
     await tester.pumpAndSettle();
 
-    expect(states, <WidgetState>{});
+    expect(delta, FWidgetStatesDelta({WidgetState.hovered}, {}));
     expect(hovered, false);
   });
 }